resetpassword

+++ app/controllers/users_controller.rb
@@ -28,11 +28,9 @@ class UsersController < ApplicationController
   end
 
   def edit
-    @user = User.find(params[:id])
   end
 
   def update
-    @user = User.find(params[:id])
     if @user.update(user_params)
       flash[:success] = "Profile updated"
       redirect_to @user
@@ -43,8 +41,12 @@ class UsersController < ApplicationController
 
   def destroy
     user = User.find(params[:id])
-    user.destroy
-    flash[:success] = "User deleted"
+    if current_user != user
+      user.destroy
+      flash[:success] = "User deleted"
+    else
+      flash[:danger] = "You can't delete yourself"
+    end
     redirect_to users_url, status: :see_other
   end
 
@@ -53,8 +55,6 @@ class UsersController < ApplicationController
   def user_params
     params.require(:user).permit(:name, :email, :password, :password_confirmation)
   end
-
-  # Before filters
 
   def logged_in_user
     unless logged_in?
